requirements.txt
# Dependencias para Eco-RVM v2.0

# Flask Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Werkzeug==3.0.1

# Validación
marshmallow==3.20.1

# Configuración
python-dotenv==1.0.0

# Vision Artificial e IA
# tensorflow>=2.15.0  # Requiere Python 3.11
opencv-python==4.9.0.80
numpy>=1.26.0
matplotlib==3.8.2

# Comunicación Serial
pyserial==3.5

# Cliente HTTP
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Linting y Formateo
black==23.12.1
flake8==7.0.0

# Producción (opcional)
gunicorn==21.2.0
